Jc/tests for default processor (#6)
* noop command

* basic tests for default processor

* change to programEndSignal

* add test for empty program

+++ hrm-emulator/src/main/java/com/github/jmcampanini/hrm/emulator/Processor.java
@@ -15,7 +15,7 @@ public interface Processor {
 
     ProgramCounter programCounter();
 
-    long steps();
+    int steps();
 
     void run(List<Command> program);
 }
